Update to 14_connect-can-rstudio-use-git.Rmd
The screenshot is outdated. 

As of RStudio v1.1.383, there is no option to  "Use Git Bash as shell for Git projects". There was an issue with opening Git Bash as the default terminal for RStudio on a Windows machine (as opposed to Command Prompt). This forum cleared up the issue: https://github.com/jennybc/happy-git-with-r/issues/67.