Replace broken link
The link to instructions on updating a fork on GitHub led to a  site in Chinese. I removed the link and added instructions. It may be worth adding that updating with this method puts you a commit ahead of the upstream repo, and in general, doesn't seem like the best way of going about things (4 clicks to create the pull request and 2 to merge???, plus switching the base is confusing.)