Update workflows-upstream-changes-into-fork.Rmd
Hi wonderful Happy Git with R maintainers!  I was recently reading Ch 31:  Get upstream changes for a fork, and the images were hard to digest because they were in a very large size with a scroll bar.  This might be improved by utilizing `knitr::include_graphics` instead, which I believe should automatically resize the image to 100%. of the screen width.  In  addition, with this function you can also add figure captions and alt text, which can be useful for those who use screen reading assistive technology. I used the chunk option `dpi=1` because that is what was playing well with my blog posts, but I am not sure if that is necessary (or works) for your set up.

I didn't think too deeply about the caption or alt text, and I only submitted this for a single figure so far.  But if you confirm that (1) you want to proceed and (2) this chunk works for this book . I am happy to help or draft the captioning and alt text for the remaining figures in this chapter.