Update connect-https-pat.Rmd
Hi - I am so thankful for this resource. 

A minor suggestion, which is to change the gitcreds::gitcreds_set() code in chapter 9 so that it can actually be copied straight into R. As of right now, the function and resulting dialogue is copied, which is a bit annoying if you're trying to paste into the R console. Please feel free to decline.